backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic[email]==2.6.3
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-multipart==0.0.9
aiofiles==23.2.1
openai-whisper==20231117
transformers==4.38.2
torch==2.2.1
spacy==3.7.4
textstat==0.7.3
vaderSentiment==3.3.2
slowapi==0.1.9
fastapi-cache2==0.2.2
loguru==0.7.2
reportlab==4.1.0
pydub==0.25.1
python-dotenv==1.0.1
svix==1.21.0
user-agents==2.2.0
python-dateutil==2.9.0
requests==2.32.3
urllib3==2.2.3
chardet==5.2.0
